code refactor making objectives a ui component
